requirements.txt
mem0ai
python-dotenv
langchain
langchain_openai
langchain-community
langchain_core
requests
opencv-python 
Pillow
pyautogui
deepface
tf-keras
langgraph
cloudinary